requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-BabelPlus
Flask-Login
Flask-WTF
Flask-Socketio
Flask-CORS
Pillow
eventlet